# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Boilerplate API",
    "version": "1.0.0",
    "description": "Enterprise-grade Express.js API documentation"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/v1/permissions/detail": {
      "post": {
        "summary": "Get details of a specific permission by its ID",
        "description": "Requires SUPER_ADMIN role. Only super admin can view permission details.",
        "tags": [
          "Permissions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "permissionId"
                ],
                "properties": {
                  "permissionId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "UUID of the permission to fetch"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Permission found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Permission fetched successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "name": {
                          "type": "string"
                        },
                        "module": {
                          "type": "string"
                        },
                        "action": {
                          "type": "string"
                        },
                        "description": {
                          "type": "string"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request (Invalid ID or not found)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Permission not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/v1/table-permissions/models": {
      "get": {
        "summary": "Get all models",
        "tags": [
          "TablePermissions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Models fetched successfully"
          }
        }
      },
      "post": {
        "summary": "Create a new model",
        "tags": [
          "TablePermissions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "modelName",
                  "tableName",
                  "module"
                ],
                "properties": {
                  "modelName": {
                    "type": "string"
                  },
                  "tableName": {
                    "type": "string"
                  },
                  "module": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Model created successfully"
          }
        }
      },
      "patch": {
        "summary": "Update a model",
        "tags": [
          "TablePermissions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "id"
                ],
                "properties": {
                  "id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "modelName": {
                    "type": "string"
                  },
                  "tableName": {
                    "type": "string"
                  },
                  "module": {
                    "type": "string"
                  },
                  "isActive": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Model updated successfully"
          }
        }
      },
      "delete": {
        "summary": "Delete a model",
        "tags": [
          "TablePermissions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Model deleted successfully"
          }
        }
      }
    },
    "/api/v1/table-permissions/models/detail": {
      "post": {
        "summary": "Get model detail by ID",
        "tags": [
          "TablePermissions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "id"
                ],
                "properties": {
                  "id": {
                    "type": "string",
                    "format": "uuid"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Model fetched successfully"
          }
        }
      }
    },
    "/api/v1/table-permissions/permissions/detail": {
      "post": {
        "summary": "Get table permissions for a model",
        "tags": [
          "TablePermissions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "modelId"
                ],
                "properties": {
                  "modelId": {
                    "type": "string",
                    "format": "uuid"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Table permissions fetched successfully"
          }
        }
      }
    },
    "/api/v1/table-permissions/permissions/upsert": {
      "post": {
        "summary": "Create or update table permissions for a model",
        "tags": [
          "TablePermissions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "modelId",
                  "permissions"
                ],
                "properties": {
                  "modelId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "permissions": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": [
                        "action"
                      ],
                      "properties": {
                        "action": {
                          "type": "string",
                          "enum": [
                            "create",
                            "read",
                            "update",
                            "delete",
                            "export",
                            "import"
                          ]
                        },
                        "scope": {
                          "type": "string",
                          "enum": [
                            "global",
                            "tenant",
                            "self",
                            "custom"
                          ],
                          "default": "global"
                        },
                        "attributes": {
                          "type": "object"
                        },
                        "abacRules": {
                          "type": "object"
                        },
                        "description": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Table permissions updated successfully"
          }
        }
      }
    },
    "/api/v1/table-permissions/permissions": {
      "patch": {
        "summary": "Update a specific table permission",
        "tags": [
          "TablePermissions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "id"
                ],
                "properties": {
                  "id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "scope": {
                    "type": "string",
                    "enum": [
                      "global",
                      "tenant",
                      "self",
                      "custom"
                    ]
                  },
                  "attributes": {
                    "type": "object"
                  },
                  "abacRules": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Table permission updated successfully"
          }
        }
      },
      "delete": {
        "summary": "Delete a table permission",
        "tags": [
          "TablePermissions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Table permission deleted successfully"
          }
        }
      }
    },
    "/api/v1/table-permissions/role-permissions/grant": {
      "post": {
        "summary": "Grant table permission to a global role",
        "tags": [
          "TablePermissions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "roleId",
                  "tablePermissionId"
                ],
                "properties": {
                  "roleId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "tablePermissionId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "expiresAt": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Permission granted successfully"
          }
        }
      }
    },
    "/api/v1/table-permissions/role-permissions/revoke": {
      "post": {
        "summary": "Revoke table permission from a global role",
        "tags": [
          "TablePermissions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "roleId",
                  "tablePermissionId"
                ],
                "properties": {
                  "roleId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "tablePermissionId": {
                    "type": "string",
                    "format": "uuid"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Permission revoked successfully"
          }
        }
      }
    },
    "/api/v1/table-permissions/role-permissions": {
      "post": {
        "summary": "Get all granted permissions for a role",
        "tags": [
          "TablePermissions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "roleId"
                ],
                "properties": {
                  "roleId": {
                    "type": "string",
                    "format": "uuid"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Role permissions fetched successfully"
          }
        }
      }
    },
...
    "/api/v1/table-permissions/tenant-role-permissions": {
      "post": {
        "summary": "Get all granted permissions for a tenant role",
        "tags": [
          "TablePermissions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "tenantRoleId"
                ],
                "properties": {
                  "tenantRoleId": {
                    "type": "string",
                    "format": "uuid"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tenant role permissions fetched successfully"
          }
        }
      }
    },
...
  },
...
}